+++ src/main/java/eggporIzquierda/solucionesactivas/SeguridadWeb.java
@@ -29,7 +29,7 @@ public class SeguridadWeb {
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
         http
                 .authorizeRequests()
-                .requestMatchers("/admin/*").hasRole("ADMIN")
+                .requestMatchers("/admin/*" , "/servicioOfrecido/*").hasRole("ADMIN")
                 .requestMatchers("/css/*", "/js/*", "/img/*", "/**")
                 .permitAll()
                 .and().formLogin()
