+++ src/main/resources/templates/dashboardadmin.html
@@ -78,8 +78,8 @@
                             <div id="collapseOne" class="accordion-collapse collapse show" data-bs-parent="#accordionExample">
                                 <div class="accordion-body">
                                     <div class="list-group">
-                                        <a th:href="@{/admin/listarServicio_ofrecido}" class="list-group-item list-group-item-action list-group-item-dark">Lista Servicios</a>
-                                        <a  th:href="@{/admin/altaservicio_ofrecido}" class="list-group-item list-group-item-action list-group-item-dark">Alta de Servicio</a>
+                                        <a th:href="@{/servicioOfrecido/listarServicio_ofrecido}" class="list-group-item list-group-item-action list-group-item-dark">Lista Servicios</a>
+                                        <a  th:href="@{/servicioOfrecido/altaservicio_ofrecido}" class="list-group-item list-group-item-action list-group-item-dark">Alta de Servicio</a>
                                     </div>     
                                 </div>
                             </div>
