+++ src/main/resources/templates/servicio_ofrecido_alta.html
@@ -10,7 +10,7 @@
         <div class="container container-fluid d-flex">
             <main class="flex-grow-1 container-fluid" id="mainindex">
 
-                     <form class="formulario form-control" th:action="@{/admin/altaservicio_ofrecido_ok}" method="POST" enctype="multipart/form-data">
+                     <form class="formulario form-control" th:action="@{/servicioOfrecido/altaservicio_ofrecido_ok}" method="POST" enctype="multipart/form-data">
                             <div class ="texto">
