+++ src/main/resources/templates/servicio_ofrecido_list.html
@@ -27,7 +27,7 @@
                                 <td th:text="${servicio.serv_descripcion}"></td>
 
                                 <td>
-                                    <a th:href= "@{/admin/modificar_servicio_ofrecido/__${servicio.serv_id}__}"><i class="fas fa-feather-alt"></i>Editar</a>
+                                    <a th:href= "@{/servicioOfrecido/modificar_servicio_ofrecido/__${servicio.serv_id}__}"><i class="fas fa-feather-alt"></i>Editar</a>
                                 </td>
                             </tr>
                         </tbody>
