+++ src/main/resources/templates/servicio_ofrecido_modificar.html
@@ -10,7 +10,7 @@
         <div class="container container-fluid d-flex">
             <main class="flex-grow-1 container-fluid" id="mainindex">
 
-                     <form class="formulario form-control" th:action="@{/admin/modificar_servicio_ofrecido/__${serv_id}__}" method="POST" enctype="multipart/form-data">
+                     <form class="formulario form-control" th:action="@{/servicioOfrecido/modificar_servicio_ofrecido/__${serv_id}__}" method="POST" enctype="multipart/form-data">
                             <div class ="texto">
 
                                 <div class=" form-input" >
